Fixed #1122 - 2.0: Eliminate Document and Properties interface to simplify the codes
