+++ shared/src/main/java/com/couchbase/lite/Properties.java
@@ -1,50 +1,238 @@
 package com.couchbase.lite;
 
+import com.couchbase.lite.internal.support.DateUtils;
+import com.couchbase.litecore.fleece.FLDict;
+import com.couchbase.litecore.fleece.FLValue;
+
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-public interface Properties extends Iterable<String> {
-    // Document properties:
+import static com.couchbase.litecore.fleece.FLConstants.FLValueType.kFLArray;
 
-    Map<String, Object> getProperties();
+class Properties {
 
-    void setProperties(Map<String, Object> properties);
+    // TODO: DB005 - sharedKeys
+    FLDict root;
+    Map<String, Object> properties;
+    // TODO: DB005 - changesKeys;
+    boolean hasChanges;
 
-    // Getter and Setter accessors:
-    Properties set(String key, Object value); // return `this`
+    //---------------------------------------------
+    // public API methods
+    //---------------------------------------------
 
-    Object get(String key); // String, Number, Boolean, SubDocument, Blob, Array(List)
+    public Map<String, Object> getProperties() {
+        if (properties == null)
+            properties = getSavedProperties();
+        return properties;
+    }
 
-    String getString(String key);
+    public void setProperties(Map<String, Object> properties) {
+        if (properties != null)
+            this.properties = new HashMap<>(properties);
+        else
+            this.properties = new HashMap<>();
 
-    int getInt(String key);
+        // TODO: DB005 - Special case handling for Blob
 
-    float getFloat(String key);
+        markChanges();
+    }
 
-    double getDouble(String key);
+    public Properties set(String key, Object value) {
+        // Date
+        if (value instanceof Date)
+            value = DateUtils.toJson((Date) value);
 
-    boolean getBoolean(String key);
+        if (hasChanges || (value == null ? getObject(key) != null : !value.equals(getObject(key)))) {
+            mutateProperties();
+            properties.put(key, value);
+            markChanges();
+        }
 
-    Blob getBlob(String key);  // TODO: DB005
+        return this;
+    }
 
-    Date getDate(String key);
+    public Object get(String key) {
+        return getObject(key);
+    }
 
-    List<Object> getArray(String key); // NOTE: Should it be List<Object> getList(String key)??
+    public String getString(String key) {
+        if (properties != null)
+            return cast(properties.get(key), String.class);
+        else {
+            FLValue flvalue = getValueFromRoot(key);
+            return flvalue == null ? null : flvalue.asString();
+        }
+    }
 
-    SubDocument getSubDocument(String key); // TODO: DB005 SubDocumentModel
+    public int getInt(String key) {
+        if (properties != null) {
+            Number obj = cast(properties.get(key), Number.class);
+            return obj == null ? 0 : obj.intValue();
+        } else {
+            FLValue flvalue = getValueFromRoot(key);
+            return flvalue == null ? 0 : flvalue.asInt();
+        }
+    }
 
-    // Relationships:
-    Document getDocument(String key);
+    public float getFloat(String key) {
+        if (properties != null) {
+            Number obj = cast(properties.get(key), Number.class);
+            return obj == null ? 0.0F : obj.floatValue();
+        } else {
+            FLValue flvalue = getValueFromRoot(key);
+            return flvalue == null ? 0.0F : flvalue.asFloat();
+        }
+    }
 
-    List<Document> getDocuments(String key); // TODO: DB00x DocumentModel
+    public double getDouble(String key) {
+        if (properties != null) {
+            Number obj = cast(properties.get(key), Number.class);
+            return obj == null ? 0.0 : obj.doubleValue();
+        } else {
+            FLValue flvalue = getValueFromRoot(key);
+            return flvalue == null ? 0.0 : flvalue.asDouble();
+        }
+    }
 
-    // Get a complex property object:
-    Property getProperty(String key); // TODO: DB00x PropertyModel
+    public boolean getBoolean(String key) {
+        if (properties != null) {
+            Object val = properties.get(key);
+            if (val == null) return false;
+            Boolean obj = cast(val, Boolean.class);
+            // NOTE: to be consistent with other platform, return true if failed to cast.
+            return obj == null ? true : obj.booleanValue();
+        } else {
+            FLValue flvalue = getValueFromRoot(key);
+            return flvalue == null ? false : flvalue.asBool();
+        }
+    }
 
-    // Remove:
-    Properties remove(String key);
+    public Blob getBlob(String key) {
+        // TODO: DB005
+        return null;
+    }
 
-    // Existence:
-    boolean contains(String key);
+    public Date getDate(String key) {
+        return DateUtils.fromJson(getString(key));
+    }
+
+    public List<Object> getArray(String key) {
+        if (properties != null) {
+            return cast(properties.get(key), List.class);
+        } else {
+            FLValue flvalue = getValueFromRoot(key);
+            // NOTE: need to check type, otherwise return empty array instead of null
+            return flvalue == null || flvalue.getType() != kFLArray ? null : flvalue.asArray();
+        }
+    }
+
+    public SubDocument getSubDocument(String key) {
+        // TODO: DB005
+        return null;
+    }
+
+    public Document getDocument(String key) {
+        // TODO: DB005
+        return null;
+    }
+
+    public List<Document> getDocuments(String key) {
+        // TODO: DB005
+        return null;
+    }
+
+    public Property getProperty(String key) {
+        // TODO: DB005
+        return null;
+    }
+
+    public Properties remove(String key) {
+        mutateProperties();
+        properties.remove(key);
+        markChanges();
+        return this;
+    }
+
+    public boolean contains(String key) {
+        if (properties != null)
+            return properties.containsKey(key);
+        else
+            // TODO: DB005 - Need to update once shared key is implemented.
+            return root == null ? false : root.asDict().containsKey(key);
+    }
+
+    //---------------------------------------------
+    // Implementation of Iterable
+    //---------------------------------------------
+
+    /**
+     * Implementing for Iterable.
+     * Currently iterator() returns keys. Not Key,Value pair.
+     *
+     * @return Iterator<String>
+     */
+    public Iterator<String> iterator() {
+        if (properties != null) {
+            return properties.keySet().iterator();
+        } else {
+            return root == null ? null : root.iterator();
+        }
+    }
+
+    //---------------------------------------------
+    // Package level access
+    //---------------------------------------------
+
+    void setHasChanges(boolean hasChanges) {
+        this.hasChanges = hasChanges;
+    }
+
+    void markChanges() {
+        if (!hasChanges)
+            hasChanges = true;
+    }
+
+    void setRoot(FLDict root) {
+        this.root = root;
+    }
+
+    //---------------------------------------------
+    // Private (in class only)
+    //---------------------------------------------
+
+    private Object getObject(String key) {
+        if (properties != null)
+            return properties.get(key);
+        else
+            return root == null ? null : root.get(key).asObject();
+    }
+
+    private void mutateProperties() {
+        if (properties == null) {
+            properties = root == null ? null : root.asDict();
+            if (properties == null)
+                properties = new HashMap<>();
+        }
+    }
+
+    private Map<String, Object> getSavedProperties() {
+        if (properties != null && !hasChanges)
+            return properties;
+        else
+            return root == null ? null : root.asDict();
+    }
+
+    private static <T> T cast(Object obj, Class<T> clazz) {
+        if (obj != null && !clazz.isInstance(obj))
+            return null;
+        return (T) obj;
+    }
+
+    private FLValue getValueFromRoot(String key) {
+        return root == null ? null : root.get(key);
+    }
 }
